Big renaming commit Generally the convention was to write `ConceptsMap` to refer to the react component.  When developing the `DetermineConceptStatusMapping` command, this posed an awkward naming convention, so `ConceptsMap` is renamed to `ConceptMap` in every file.  This includes in code (controllers, tests, helpers, javascript) and also files/directories.
All rails tests pass
All jest tests pass
