+++ app/commands/user_track/generate_concept_status_mapping.rb
@@ -1,16 +1,10 @@
 class UserTrack
-  class GenerateConceptsStatus
+  class GenerateConceptStatusMapping
     include Mandate
 
     initialize_with :user_track
 
     def call
-      concepts_status
-    end
-
-    private
-    memoize
-    def concepts_status
       all_concepts = user_track.track.concepts.map(&:slug)
       available_concepts = user_track.available_concepts.map(&:slug)
       learnt_concepts = user_track.learnt_concepts.map(&:slug)
