+++ app/controllers/tracks/concepts_controller.rb
@@ -7,7 +7,7 @@ class Tracks::ConceptsController < ApplicationController
 
   def authenticated_index
     if current_user.joined_track?(@track)
-      @layout_data = Track::DetermineConceptsLayout.(@track)
+      @layout_data = Track::DetermineConceptMapLayout.(@track)
       render action: "index/joined"
     else
       render action: "index/unjoined"
