+++ app/javascript/components/concept-map/CompleteIcon.tsx
