+++ app/javascript/components/concept-map/Concept.tsx
