+++ app/javascript/components/concept-map/ConceptMap.css
