+++ app/javascript/components/concept-map/concept-connection-types.ts
