+++ app/javascript/components/concept-map/concept-map-types.ts
