+++ app/javascript/components/concept-map/concept-types.ts
