+++ app/javascript/components/concept-map/helpers/draw-helpers.ts
