+++ app/javascript/components/concept-map/helpers/path-helpers.ts
