+++ app/javascript/components/concept-map/hooks/useWebpageSize.ts
