+++ app/javascript/packs/application.tsx
@@ -13,7 +13,7 @@ import 'css/styles.css'
 import 'css/layout.css'
 import 'css/components/tracks-list.css'
 import 'css/components/track.css'
-import 'components/concepts-map/ConceptsMap.css'
+import 'components/concept-map/ConceptMap.css'
 
 import React from 'react'
 import { initReact } from './react-bootloader.jsx'
@@ -22,7 +22,7 @@ import * as Maintaining from '../components/maintaining'
 import * as Notifications from '../components/notifications'
 import * as Mentoring from '../components/mentoring'
 import * as Student from '../components/student'
-import { ConceptsMap } from '../components/concepts-map/ConceptsMap'
+import { ConceptMap } from '../components/concept-map/ConceptMap'
 
 // Add all react components here.
 // Each should map 1-1 to a component in app/helpers/components
@@ -56,8 +56,8 @@ initReact({
       tagOptions={data.tag_options}
     />
   ),
-  'concepts-map': (data: any) => (
-    <ConceptsMap
+  'concept-map': (data: any) => (
+    <ConceptMap
       concepts={data.graph.concepts}
       levels={data.graph.levels}
       connections={data.graph.connections}
