+++ app/views/test/components/student/concept_map/show.html.haml
@@ -1,0 +1,1 @@
+= ViewComponents::Student::ConceptMap.new(@data)
