--- app/views/test/components/student/concepts_map/show.html.haml
@@ -1,1 +1,0 @@
-= ViewComponents::Student::ConceptsMap.new(@data)
