+++ app/views/tracks/concepts/index/joined.html.haml
@@ -1,3 +1,3 @@
 .md-container
   %h2 Concepts for #{@track.title}
-  = ViewComponents::Student::ConceptsMap.new(@layout_data)
+  = ViewComponents::Student::ConceptMap.new(@layout_data)
