+++ test/helpers/view_components/student/concept_map_test.rb
@@ -1,16 +1,16 @@
 require_relative "../view_component_test_case"
 
 module Student
-  class ConceptsMapTest < ViewComponentTestCase
+  class ConceptMapTest < ViewComponentTestCase
     test "component with empty concepts map rendered correctly" do
       data = {
         concepts: [],
         levels: [],
         connections: []
       }
-      component = ViewComponents::Student::ConceptsMap.new(data).to_s
+      component = ViewComponents::Student::ConceptMap.new(data).to_s
 
-      assert_component component, "concepts-map", { graph: data }
+      assert_component component, "concept-map", { graph: data }
     end
   end
 end
